* upstream/master: (868 commits)
  Add utility functions for interpreting ThreadStatus
  rocksdb: Enable scan-build static analysis
  Remove unused util/thread_event_info.h
  Changes for supporting cross functional tests for inplace_update
  Divide test DBIteratorTest.DBIterator to smaller tests
  Options::PrepareForBulkLoad() to increase parallelism of flushes
  Fix unity build
  NewIteratorWithBase() for default column family
  Fix compile warning in util/xfunc.h
  Cross functional test infrastructure for RocksDB.
  remove old debugging message (#487)
  Release.md - Remove version change instrcution
  Update HISTORY-JAVA.md
  [RocksJava] Deprecate setSkipLogErrorOnRecovery
  [RocksJava] Snapshot - GetSequenceNumber
  [RocksJava] Incorporated changes for D32151
  [RocksJava] GetUpdatesSince support
  [RocksJava] Adjusted auto pointer
  [RocksJava] enable/disable File deletions
  [RocksJava] GetUpdatesSince support
  ...

+++ table/block_hash_index.cc
@@ -59,7 +59,7 @@ BlockHashIndex* CreateBlockHashIndexOnTheFly(
   auto hash_index = new BlockHashIndex(
       hash_key_extractor,
       true /* hash_index will copy prefix when Add() is called */);
-  uint64_t current_restart_index = 0;
+  uint32_t current_restart_index = 0;
 
   std::string pending_entry_prefix;
   // pending_block_num == 0 also implies there is no entry inserted at all.
